LayoutTests/fast/text/variations/getComputedStyle.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<style>
@font-face {
	font-family: "Boxis";
	src: url("resources/Boxis-VF.ttf");
}
</style>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="test0" style="font-family: 'Boxis'; font-variation-settings: 'hght' 400;">a</div>
<div id="test1" style="font-family: 'Boxis'; font-variation-settings: 'hght' 400, 'hght' 500;">a</div>
<div id="test2" style="font-family: 'Boxis'; font-variation-settings: 'hght' 400, 'wdth' 500;">a</div>
<div id="test3" style="font-family: 'Boxis'; font-variation-settings: 'wdth' 500, 'hght' 400;">a</div>
<div id="test4" style="font-family: 'Boxis'; font-variation-settings: normal;">a</div>
<div id="test5" style="font-family: 'Boxis'; font-variation-settings: 'abcde' 500;">a</div>
<div id="test6" style="font-family: 'Boxis'; font-variation-settings: 'abc한국어e' 500;">a</div>
<div id="test7" style="font-family: 'Boxis'; font-variation-settings: 'hght' 999999;">a</div>
<script>
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test0')).getPropertyValue('font-variation-settings')", "\"hght\" 400");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).getPropertyValue('font-variation-settings')", "\"hght\" 500");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-variation-settings')", "\"wdth\" 500, \"hght\" 400");
shouldBe("window.getComputedStyle(document.getElementById('test3')).getPropertyValue('font-variation-settings')", "window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-variation-settings')");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).getPropertyValue('font-variation-settings')", "normal");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).getPropertyValue('font-variation-settings')", "normal");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).getPropertyValue('font-variation-settings')", "normal");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test7')).getPropertyValue('font-variation-settings')", "\"hght\" 999999");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
